pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "marketgemini-backend"
version = "0.1.0"
description = "MarketGemini backend (FastAPI)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0",
  "pyjwt[crypto]", 
  "httpx>=0.27",
  "PyJWT>=2.9",
  "itsdangerous>=2.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["marketgemini_backend*"]